Pagination and simple search added

+++ src/main/java/uz/aytjanov/googlephotosclone/controller/PhotosController.java
@@ -1,17 +1,19 @@
 package uz.aytjanov.googlephotosclone.controller;
 
 import jakarta.servlet.http.HttpSession;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.*;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
-import uz.aytjanov.googlephotosclone.dto.PhotoListDto;
 import uz.aytjanov.googlephotosclone.dto.ResponseDto;
 import uz.aytjanov.googlephotosclone.entity.Photo;
 import uz.aytjanov.googlephotosclone.service.PhotosService;
 import uz.aytjanov.googlephotosclone.service.SuperbaseStorageService;
 import java.io.IOException;
-import java.util.List;
 import java.util.Map;
 
 @RestController
@@ -24,30 +26,38 @@ public class PhotosController {
         this.superbaseStorageService = superbaseStorageService;
     }
 
-    private Long requireUserId(HttpSession session) {
+    private boolean checkAuthSession(HttpSession session) {
         Long userId = (Long) session.getAttribute("userId");
-        if (userId == null) {
-
-            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
-        }
-        return userId;
+        return !(userId == null);
+    }
+    private Long getUserId(HttpSession session) {
+        return (Long) session.getAttribute("userId");
     }
     private boolean isPhotoBelongsToUser(Photo photo, Long userId) {
         if (photo == null) return false;
         return photo.getUser().getId().equals(userId);
     }
    @GetMapping("/api/photos")
-   public ResponseEntity<?> photos(HttpSession session) {
-        Long userId = requireUserId(session);
-        List<PhotoListDto> result = photosService.getAllUsersPhotosAsDtos(userId);
-        return ResponseEntity.ok(result);
+   public ResponseEntity<Page<Photo>> photos(HttpSession session,
+                                             @RequestParam(defaultValue = "")
+                                             String search, @RequestParam(defaultValue = "0")
+                                             int page, @RequestParam(defaultValue = "12")
+                                             int size) {
+       if (checkAuthSession(session)) {
+           Pageable pageable = PageRequest.of(page, size, Sort.by("id").descending());
+           return ResponseEntity.ok(photosService.searchPhoto(search, pageable));
+       }
+       return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
    }
    @GetMapping("/api/photos/{id}")
    public ResponseEntity<byte[]> openFile(@PathVariable Long id, HttpSession session) throws IOException {
-        Long userId = requireUserId(session);
-        Photo photo = photosService.getPhoto(id);
-        if (!isPhotoBelongsToUser(photo, userId)) return ResponseEntity.notFound().build();
-        return photosService.openTheFile(photo);
+        if (checkAuthSession(session)) {
+            Long userId = getUserId(session);
+            Photo photo = photosService.getPhoto(id);
+            if (!isPhotoBelongsToUser(photo, userId)) return ResponseEntity.notFound().build();
+            return photosService.openTheFile(photo);
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
    }
    @PostMapping("/api/logout")
    public ResponseEntity<?> logout(HttpSession session) {
@@ -60,22 +70,31 @@ public class PhotosController {
         if (file.isEmpty()) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "File is empty!");
         }
-        Long userId = requireUserId(session);
-        Photo photo = superbaseStorageService.uploadFile(userId, file);
-        return ResponseEntity.status(HttpStatus.CREATED).body(new ResponseDto(photo.getFileName(), photo.getContentType()));
+        if (checkAuthSession(session)) {
+            Long userId = getUserId(session);
+            Photo photo = superbaseStorageService.uploadFile(userId, file);
+            return ResponseEntity.status(HttpStatus.CREATED).body(new ResponseDto(photo.getFileName(), photo.getContentType()));
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @DeleteMapping("/api/photos/{id}")
     public ResponseEntity<HttpStatus> delete(@PathVariable Long id, HttpSession session) {
-        Long userId = requireUserId(session);
-        Photo photo = photosService.getPhoto(id);
-        if (!photo.getUser().getId().equals(userId)) throw new ResponseStatusException(HttpStatus.FORBIDDEN);
-        photosService.delete(id);
-        return ResponseEntity.noContent().build();
+        if (checkAuthSession(session)) {
+            Long userId = getUserId(session);
+            Photo photo = photosService.getPhoto(id);
+            if (!photo.getUser().getId().equals(userId)) throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+            photosService.delete(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
     @GetMapping("/api/photos/{id}/download")
     public ResponseEntity<byte[]> download(@PathVariable Long id, HttpSession session) throws IOException {
-        Long userId = requireUserId(session);
-        return photosService.download(id, userId);
+        if (checkAuthSession(session)) {
+            Long userId = getUserId(session);
+            return photosService.download(id, userId);
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 }
